Extract some helpers from navigation.js
